+++ components/EntryPage/EntryPageWrap/EntryPageWrap.styled.ts
@@ -1,21 +1,12 @@
 import styled from 'styled-components';
 import { colors } from '@/variables/colors.styled';
-import { fonts } from '@/variables/fonts.styled';
 
-export const PageWrap = styled.span`
-  height: 100vh;
+export const PageWrap = styled.div`
+  min-height: 95vh;
   width: 100%;
   display: flex;
   flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  background-color: ${colors.colorBgSecondaryDarken};
-`
-
-export const ContentWrap = styled.div`
-  height: 100%;
-  width: 100%;
-  display: flex;
   justify-content: center;
   align-items: center;
+  background-color: ${colors.colorBgSecondaryDarken};
 `
