+++ components/Footer/Footer.tsx
@@ -4,7 +4,6 @@ import {
   FooterWrapper,
   Link,
   GitHubsList,
-  GitHubsItem,
   LogoSchool,
   Wrap,
 } from './footer.styled';
@@ -17,15 +16,15 @@ export const Footer: React.FC = () => {
           <LogoSchool src="/rs_school_js.svg" alt="rsschool" />
         </Link>
         <GitHubsList>
-          <GitHubsItem>
+          <li>
             <Link href="https://github.com/Voloshin-Sergei">@Voloshin-Sergei</Link>
-          </GitHubsItem>
-          <GitHubsItem>
+          </li>
+          <li>
             <Link href="https://github.com/Maslovars">@Maslovars</Link>
-          </GitHubsItem>
-          <GitHubsItem>
+          </li>
+          <li>
             <Link href="https://github.com/YuliaEnik">@YuliaEnik </Link>
-          </GitHubsItem>
+          </li>
         </GitHubsList>
         <Wrap>
           <Link href="https://rs.school/react/">REACT </Link>
