+++ components/Footer/footer.styled.ts
@@ -3,12 +3,13 @@ import { colors } from '@/variables/colors.styled';
 import { fonts } from '@/variables/fonts.styled';
 import { breakpoints } from '@/variables/response.styled';
 
-export const FooterSpan = styled.span`
+export const FooterSpan = styled.div`
   width: 100%;
   background-color: ${colors.colorBgSecondary};
 `;
 
 export const FooterWrapper = styled.div`
+  min-height: 5vh;
   display: flex;
   flex-wrap: wrap;
   gap: 1rem;
@@ -17,10 +18,11 @@ export const FooterWrapper = styled.div`
   max-width: 140rem;
   height: 100%;
   margin: 0 auto;
-  padding: 0.5rem 2rem;
   font-size: ${fonts.fontSizeRegular};
   @media (max-width: ${breakpoints.sm}) {
     justify-content: center;
+    gap: .5rem;
+    font-size: ${fonts.fontSizeRegular}*.6;
   }
 `;
 
@@ -34,8 +36,12 @@ export const Link = styled.a`
 `;
 
 export const LogoSchool = styled.img`
-  width: 80px;
-  height: 40px;
+  width: 8rem;
+  height: 4rem;
+  @media (max-width: ${breakpoints.sm}) {
+    width: 6rem;
+    height: 3rem;  
+  }
 `;
 
 export const GitHubsList = styled.ul`
@@ -43,11 +49,18 @@ export const GitHubsList = styled.ul`
   flex-wrap: wrap;
   gap: 1rem;
   justify-content: center;
+  @media (max-width: ${breakpoints.sm}) {
+    gap: .2rem;
+    order: 1;
+  }
 `;
 
-export const GitHubsItem = styled.li``;
 export const Wrap = styled.div`
   display: flex;
   flex-wrap: wrap;
-  gap: 1rem;
+  column-gap: 1rem;
+  @media (max-width: ${breakpoints.sm}) {
+    gap: .2rem;
+ 
+  }
 `;
