+++ components/Layout/Layout.tsx
@@ -1,0 +1,11 @@
+type Props = {
+  children?: React.ReactNode,
+};
+
+export const Layout: React.FC<Props> = ({children}) => {
+  return (
+    <>
+      {children}
+    </>
+  )
+};
