+++ pages/_app.tsx
@@ -1,11 +1,16 @@
 import type { AppProps } from 'next/app';
-import { GlobalStyle } from '../styles/global.styled';
+import { GlobalStyle } from 'styles/global.styled';
+import { Layout } from '@/components/Layout/Layout';
+import { Footer } from '@/components/Footer';
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
     <>
-      <GlobalStyle />
-      <Component {...pageProps} />
+      <Layout>
+        <GlobalStyle />
+        <Component {...pageProps} />
+      </Layout>
+      <Footer />
     </>
   );
 }
