--- pages/signup/SignUp.styled.ts
@@ -1,2 +1,0 @@
-import styled from 'styled-components';
-import { colors } from '@/variables/colors.styled';
