+++ pages/welcome/welcome.styled.ts
@@ -8,8 +8,8 @@ const WelcomeSpan = styled.div`
   flex-direction: column;
   align-items: center;
   justify-content: space-between;
-  width: 100vw;
-  height: 100vh;
+  width: 100%;
+  height: 95vh;
   background-color: ${colors.colorBgPrimaryLighten};
 `;
